+++ src/extension.ts
@@ -1,9 +1,10 @@
 import * as vscode from "vscode";
 import { Command } from "./config";
-import { ProblemTreeDataProvider, ProblemTreeDecorationProvider, ProblemViewProvider } from "./view";
+import { ProblemDataTreeItem, ProblemTreeDataProvider, ProblemTreeDecorationProvider, ProblemViewProvider } from "./view";
 import { ProblemData, ProblemDataService } from "./service";
 import { EulerLensProvider } from "./lens";
 import { EulerResourceProvider } from "./resource";
+import { FavoritesDataProvider } from "./view/favorites";
 
 interface SearchItem extends vscode.QuickPickItem {
     id: number,
@@ -13,7 +14,7 @@ interface SearchItem extends vscode.QuickPickItem {
 
 async function search(this: ProblemDataService) {
     const problemData: ProblemData[] = this.getProblemInfo();
-    const items = problemData.map((data) => {
+    const items = problemData.filter((data) => data !== null).map((data) => {
         return {
             id: data.ID,
             label: `Problem ${data.ID}`,
@@ -28,7 +29,24 @@ async function search(this: ProblemDataService) {
             matchOnDescription: true,
             matchOnDetail: true,
         }).then((selected) => { if (selected) { vscode.commands.executeCommand(Command.Show, selected?.id) } });
+}
 
+async function newSubFolder(this: FavoritesDataProvider, target?: ProblemDataTreeItem) {
+    let folderName = await vscode.window.showInputBox({
+        prompt: "Enter folder name"
+    });
+    if (folderName) {
+        this.newFolder(folderName, target)
+    }
+}
+
+async function newFolder(this: FavoritesDataProvider) {
+    let folderName = await vscode.window.showInputBox({
+        prompt: "Enter folder name"
+    });
+    if (folderName) {
+        this.newFolder(folderName)
+    }
 }
 
 export function activate(context: vscode.ExtensionContext) {
@@ -37,6 +55,19 @@ export function activate(context: vscode.ExtensionContext) {
     const viewProvider = new ProblemViewProvider(context, dataService);
     const resourceProvider = new EulerResourceProvider(context);
     const treeProvider = new ProblemTreeDataProvider(dataService, context);
+
+    const view = vscode.window.createTreeView("projecteuler.problemView", {
+        treeDataProvider: treeProvider,
+        canSelectMany: true,
+        dragAndDropController: treeProvider
+    });
+
+    const favoritesProvider = new FavoritesDataProvider(dataService, context);
+    const favorites = vscode.window.createTreeView("projecteuler.favorites", {
+        treeDataProvider: favoritesProvider,
+        canSelectMany: false,
+        dragAndDropController: favoritesProvider
+    });
     context.subscriptions.push(
         vscode.workspace.registerTextDocumentContentProvider("resource", resourceProvider),
         vscode.commands.registerCommand(Command.Search, search, dataService),
@@ -44,11 +75,14 @@ export function activate(context: vscode.ExtensionContext) {
         vscode.commands.registerCommand(Command.Show, viewProvider.showProblem, viewProvider),
         vscode.commands.registerCommand(Command.Refresh, dataService.refreshMetadata, dataService),
         vscode.commands.registerCommand(Command.Clear, dataService.clearMetadata, dataService),
+        vscode.commands.registerCommand(Command.NewFolder, newFolder, favoritesProvider),
+        vscode.commands.registerCommand(Command.NewSubFolder, newSubFolder, favoritesProvider),
+        vscode.commands.registerCommand(Command.AddToFavorites, favoritesProvider.addProblemTreeItem, favoritesProvider),
+        vscode.commands.registerCommand(Command.Delete, favoritesProvider.deleteProblemTreeItem, favoritesProvider),
+        vscode.window.registerFileDecorationProvider(new ProblemTreeDecorationProvider()),
+        view,
+        favorites,
     );
-
-    vscode.window.registerTreeDataProvider("projecteuler.problemView", treeProvider);
-    vscode.window.registerFileDecorationProvider(new ProblemTreeDecorationProvider());
-    dataService.init();
 }
 
 // This method is called when your extension is deactivated